+++ desktop/php/speedtestByOokla.php
@@ -131,7 +131,7 @@ $eqLogics = eqLogic::byType($plugin->getId());
 							<!-- La fonction cron de la classe du plugin doit contenir le code prévu pour que ce champ soit fonctionnel -->
 							<div class="form-group">
 								<label class="col-sm-4 control-label">{{Auto-actualisation}}
-									<sup><i class="fas fa-question-circle tooltips" title="{{Fréquence de rafraîchissement des commandes infos de l'équipement et de la liste des serveurs les plus proches}}"></i></sup>
+									<sup><i class="fas fa-question-circle tooltips" title="{{Permet de définir la fréquence de rafraîchissement des commandes infos et de la liste des serveurs les plus proches}}"></i></sup>
 								</label>
 								<div class="col-sm-6">
 									<div class="input-group">
@@ -145,8 +145,19 @@ $eqLogics = eqLogic::byType($plugin->getId());
 								</div>
 							</div>
 							<div class="form-group">
+								<label class="col-sm-4 control-label">{{Templates}}
+									<sup><i class="fas fa-question-circle tooltips" title="{{Permet d'utiliser les templates dédiés au plugin à la place des widgets core (paramètres dans Configuration avancée de l'équipement => onglet Affichage => section Widget}}"></i></sup>
+								</label>
+								<div class="col-sm-6">
+									<select class="eqLogicAttr form-control" data-l1key="configuration" data-l2key="template">
+										<option value="coreWidget">Widgets core</option>
+										<option value="templateWithoutGauges">Template sans jauges</option>
+									</select>
+								</div>
+							</div>
+							<div class="form-group">
 								<label class="col-sm-4 control-label">{{ID du serveur de test}}
-									<sup><i class="fas fa-question-circle tooltips" title="{{A utiliser uniquement si le serveur choisi automatiquement retourne des valeurs erronées}}"></i></sup>
+									<sup><i class="fas fa-question-circle tooltips" title="{{Permet de choisir un serveur de test spécifique (à utiliser uniquement si le serveur choisi automatiquement retourne des valeurs erronées)}}"></i></sup>
 								</label>
 								<div class="col-sm-6">
 									<input type="number" class="eqLogicAttr form-control" data-l1key="configuration" data-l2key="serverId">
@@ -154,7 +165,7 @@ $eqLogics = eqLogic::byType($plugin->getId());
 							</div>
 							<div class="form-group">
 								<label class="col-sm-4 control-label">{{Désactiver les erreurs}}
-									<sup><i class="fas fa-question-circle tooltips" title="{{Cocher cette case pour désactiver les erreurs lorsque votre connexion internet est hors service}}"></i></sup>
+									<sup><i class="fas fa-question-circle tooltips" title="{{Permet de désactiver les erreurs lorsque votre connexion internet est hors service}}"></i></sup>
 								</label>
 								<div class="col-sm-6">
 									<input type="checkbox" class="eqLogicAttr" data-l1key="configuration" data-l2key="disableError">
@@ -167,7 +178,7 @@ $eqLogics = eqLogic::byType($plugin->getId());
 						<div class="col-lg-6">
 							<legend><i class="fas fa-info"></i> {{Informations}}</legend>
 							<div class="form-group">
-								<label class="col-sm-4 control-label">{{Description}}</label>
+								<label class="col-sm-4 control-label">{{Description de l’équipement}}</label>
 								<div class="col-sm-6">
 									<textarea class="form-control eqLogicAttr autogrow" data-l1key="comment"></textarea>
 								</div>
@@ -175,7 +186,6 @@ $eqLogics = eqLogic::byType($plugin->getId());
 							<legend><i class="fas fa-server"></i> {{Liste des serveurs les plus proches}}</legend>
 							<div class="form-group">
 								<div class="alert alert-info col-xs-10 col-xs-offset-1">
-									<!-- <pre class="eqLogicAttr" data-l1key="configuration" data-l2key="serverList" style="font-family: CamingoCode, monospace;"></pre> -->
 									<pre class="serverList" style="font-family: CamingoCode, monospace;"></pre>
 								</div>
 							</div>