+++ docs/fr_FR/beta/changelog.md
@@ -4,6 +4,14 @@
 >
 > S'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte.
 
+## 12/11/2022
+
+- Ajout d'un premier template dédié au plugin version Dashboard (paramètres dans Configuration avancée de l'équipement => onglet Affichage => section Widget)
+
+## 09/11/2022
+
+- Première version stable
+
 ## 07/11/2022
 
 - Ajout d'un cron dédié au plugin dans le moteur de tâches